ENH: Added "Update rate" option to be able to control volume rendering quality/speed
In virtual reality views, interactive frame rate that is set in Slicer Volume Rendering settings was not used. Therefore, there was no way of adaptively setting volume rendering quality.

The new DesiredUpdateRate option in vtkMRMLVirtualRealityViewNode controls the desired update rate in the VR render window. If volume rendering quality is set to adaptive then this update rate is used for determining display quality.