ENH: Add new interactor class to make gesture recognition more fluid
The new mechanism instead of the VTK default pan/rotate/zoom gestures allows grabbing two "air" points in the scene, and as the controllers move, the PhysicalToWorld matrix is modified so that the controller physical pose corresponds to the same rendering world pose. It is found much more intuitive than three separate gestures.

This change overrides RecognizeComplexGesture, so that also the overhead and delay of recognizing the gesture is removed. As there is only one gesture (we call it Pinch3D), it is now possible enter into that interaction as soon as both grab buttons are pressed.

Minor changes:
- Replace VTK_OVERRIDE with override everywhere
- Fix bug that magnification did not work until the first controller interaction (this->InteractorStyle->SetCurrentRenderer(this->Renderer); in ViewPrivate::createRenderWindow)