ENH: Add optimize scene for VR button on toolbar
The optimize scene for VR button on the toolbar is a one-time action that changes some MRML properties to make VR experience better. So far:
- Hide model and segmentation slice intersections to speed up rendering (cutting is slow)
- Turn off backface culling for models so that the back side is visible when the user goes "inside" a model
- Apply these in the default display nodes so that new nodes inherit these properties

The commit also fixes a bug, which caused problems with prop picking. If a non-selectable node was in the pick position and the displayable manager was registered after the one with selectable prop in the same position, then positioning the prop did not happen. Now the picked node is only set in the first place if it is selectable.