BUG: Fix loading scene with VR; Fix segmentation slice intersection
Optimize scene did not disable slice intersection for newly created segmentations, now it is fixed.

When a scene was loaded in which a VR node was present (with potentially changed properties) then it was ignored. By using the existing VR view node instead of always creating a new one when connecting, this is averted.