fix #30: Added ControllerModelsVisible property to vtkMRMLVirtualRealityViewNode
Also added a checkbox on virtual reality module GUI.

This allows replacing default controller models by any models in the scene (by enabling ControllerTransformsUpdate and applying the transforms to models). These models are defined in world and not coordinate system, so their size is changed if world size is changed. This may or may not be desirable, so in the future methods could be added to allow defining controller models in physical coordinate system.