ENH: Make trigger button configurable
Some users find it counterintuitive that the grip button does the basic functions while the trigger button does nothing. That change was made so that the trigger button is left for the laser and UI manipulation. Since those developments are delayed significantly, it makes sense to add back the "grab objects and world" function back to the trigger button as well.

It is done in a configurable way. The interactor style has a new SetTriggerButtonFunction method that takes a string. If the string is empty then the trigger button will not do anything. If it is the constant accessible with GetButtonFunctionIdForGrabObjectsAndWorld, then it will be the same as what the grip button currently does as well. The latter is the default for now.