COMP: Configure VTKRenderingOpenVR external project passing PYTHON_* vars
This is required so that the call to to FindPackage(Python) find
the Slicer python library and include directory instead of the one
available on the system.

Note that a better solution would be to improve VTK build system by
ensuring externally built module use the same python library that was
used to configure VTK (e.g by adding them to VTKConfig). Since at the
time of this commit, the upstream VTK build system has been written and
was different from the older one still used in Slicer/VTK fork, this
approach didn't make sense.

Fixes #64

Co-authored-by: Jean-Christophe Fillion-Robin <jchris.fillionr@kitware.com>