BUG: Improve action manifest paths
The path to the main action manifest file missed a path separator character ('\') between the last folder name and the file name. This commit adds it.
The json files still need to be copied to the vtkRenderingOpenVR-build folder from vtkRenderingOpenVR-build\externals\vtkRenderingOpenVR, where they are found in the build tree.
Similar issue with the installer.

Re #115