BUG: Restore handling of complex gesture and fix event processing
This commit revisits the invocation of `vtkMRMLViewInteractorStyle::ProcessEvents`
to prioritize event processing specific to `VirtualReality`. The additional
processing is now performed only if the events have not been processed already.

Furthermore, this commit restores the support for calling `OnStartGesture`
and `OnEndGesture`, which was inadvertently omitted in the previous commit
titled "BUG: Fix integration with updated Slicer event delegation and VTK
OpenVR API."

Co-authored-by: Sankhesh Jhaveri <sankhesh.jhaveri@kitware.com>