ENH: Improve initialization process robustness using `GetVRInitialized()`
Since having the head-mounted display (HMD) connected may not always be
a good proxy to check if the initialization succeeded, we instead swith to
the dedicated `GetVRInitialized` function available since commit
kitware/VTK@bddd94efd3 ("Improve OpenVR/OpenXR module and test and enable
OpenXR in CI", 2023-01-11)

Moreover, the determination of whether the event loop should run has been
refined by relying on the GetVRInitialized function as the primary criterion.
In the case of the OpenVR runtime, the use of GetHMD is retained through the
introduction of the `hmdConnected` internal variable for enhanced flexibility.