ENH: Add OpenXR Remoting support
This commit updates the view node by adding "Remoting" and "PlayerIPAddress"
properties. The corresponding UI updates include a checkbox for toggling remoting
and a `QLineEdit` for entering the IP address.

The macro `SlicerVirtualReality_HAS_OPENXRREMOTING_SUPPORT` is added to
`vtkMRMLVirtualRealityConfigure.h.in` to conditionally include code specific
to OpenXR remoting.

The "Remoting" checkbox and "PlayerIPAddress" line edit are disabled if an
active connection has been established.

To fix OpenXRRemoting "wglDXRegisterObjectNV failed in RegisterSharedTexture()",
the OpenXRRemoting HelperWindow MultiSamples property is explicitly set to 0.

To ensure no background is displayed, the background alpha/color/gradient is
explicitly set when remoting is used.

The last "Remoting" and "PlayerIPAddress" properties associated with a successful
hardware connection are saved to and restored from the settings.

Co-authored-by: Lucas Gandel <lucas.gandel@kitware.com>